Initial setup

+++ InvokePowerShellNotebook.ps1
@@ -15,53 +15,65 @@ West   melon 76
     param(
         [Parameter(ValueFromPipelineByPropertyName)]
         $NoteBookFullName,
+        $OutFile,
         [Switch]$AsExcel,
         [Switch]$Show
     )
 
     Process {
-        $codeBlocks = @(Get-NotebookContent $NoteBookFullName -JustCode)
-        $codeBlockCount = $codeBlocks.Count
-        $SheetCount = 0
+        
+        if ($OutFile) {
+            if (Test-Path $OutFile) {
+                throw "$OutFile already exists"
+            }
 
-        for ($idx = 0; $idx -lt $codeBlockCount; $idx++) {
+            $null = Copy-Item $NoteBookFullName $OutFile
+        }
+        else {
 
-            $targetCode = $codeblocks[$idx].source
+            $codeBlocks = @(Get-NotebookContent $NoteBookFullName -JustCode)
+            $codeBlockCount = $codeBlocks.Count
+            $SheetCount = 0
 
-            Write-Progress -Activity "Executing PowerShell code block - [$(Get-Date)]" -Status (-join $targetCode) -PercentComplete (($idx + 1) / $codeBlockCount * 100)
+            for ($idx = 0; $idx -lt $codeBlockCount; $idx++) {
 
-            if ($AsExcel) {
-                if (!(Get-Module -ListAvailable ImportExcel -ErrorAction SilentlyContinue)) {
-                    throw "This feature requires the ImportExcel PowerShell module. Use 'Install-Module -Name ImportExcel' to get it from the PS Gallery."
-                }
+                $targetCode = $codeblocks[$idx].source
 
-                if ($idx -eq 0) {
-                    $notebookFileName = Split-Path $NoteBookFullName -Leaf
-                    $xlFileName = $notebookFileName -replace ".ipynb", ".xlsx"
+                Write-Progress -Activity "Executing PowerShell code block - [$(Get-Date)]" -Status (-join $targetCode) -PercentComplete (($idx + 1) / $codeBlockCount * 100)
 
-                    $xlfile = "{0}\{1}" -f $pwd.Path, $xlFileName
-                    Remove-Item $xlfile -ErrorAction SilentlyContinue
-                }
+                if ($AsExcel) {
+                    if (!(Get-Module -ListAvailable ImportExcel -ErrorAction SilentlyContinue)) {
+                        throw "This feature requires the ImportExcel PowerShell module. Use 'Install-Module -Name ImportExcel' to get it from the PS Gallery."
+                    }
 
-                foreach ($dataSet in , @($targetCode | Invoke-Expression)) {
-                    if ($dataSet) {
-                        $SheetCount++
-                        $uniqueName = "Sheet$($SheetCount)"
-                        Export-Excel -InputObject $dataSet -Path $xlfile -WorksheetName $uniqueName -AutoSize -TableName $uniqueName
+                    if ($idx -eq 0) {
+                        $notebookFileName = Split-Path $NoteBookFullName -Leaf
+                        $xlFileName = $notebookFileName -replace ".ipynb", ".xlsx"
+
+                        $xlfile = "{0}\{1}" -f $pwd.Path, $xlFileName
+                        Remove-Item $xlfile -ErrorAction SilentlyContinue
+                    }
+
+                    foreach ($dataSet in , @($targetCode | Invoke-Expression)) {
+                        if ($dataSet) {
+                            $SheetCount++
+                            $uniqueName = "Sheet$($SheetCount)"
+                            Export-Excel -InputObject $dataSet -Path $xlfile -WorksheetName $uniqueName -AutoSize -TableName $uniqueName
+                        }
                     }
                 }
+                else {
+                    , @($targetCode | Invoke-Expression)
+                }
             }
-            else {
-                , @($targetCode | Invoke-Expression)
-            }
-        }
 
-        if ($AsExcel) {
-            if ($Show) {
-                Invoke-Item $xlfile
-            }
-            else {
-                $xlfile
+            if ($AsExcel) {
+                if ($Show) {
+                    Invoke-Item $xlfile
+                }
+                else {
+                    $xlfile
+                }
             }
         }
     }
